provider: carve gemini.rs (pure move)
The Gemini region moves verbatim out of provider/mod.rs:
generateContent request assembly, the ChatProvider impl, and CU
function parsing, plus the 7 tests that exercise them. Moved items
widen to pub(crate); no cross-boundary field access this time.
